Fix specular transmission and dielectric fresnel components
It seems there are some conflicting expectations about which way `w_o`
is facing relative to the surface (eg. away or towards) throughout the
specular transmission & fresnel component code which causes some odd bugs
with specular glass objects. The fix right now is a bit hacky in that
it just patches over these inconsistencies but a proper fix would be to
derive the correct equations and pick a consisten understanding of
which way `w_o` faces across the specular and fresnel components.