Finally decided how I want to handle emissive instances
We'll have a Receiver and Emitter struct for these types where the
Emitter struct will require its geometry is also Sampleable. Then
Instance becomes an enum with Emitter and Receiver variants that simply
hold an Emitter or Receiver struct. To accelerate finding of lights for
direct light sampling each thread will have a small vec of references to
each emissive instance so we don't need a vec of arcs. Although a
Vec<Arc<Emitter>> might also work fine to have it shared with all the
threads.

Additionally this lets us unify point lights with area lights in that we
can make the Emitter's geom member an option and if it has no geometry
it will just behave as a point light. Then the emitter will also
implement the Light trait so we can use it as a light source.