Thinking about how to setup emitter to handle point & area lights
I'm not really sure if I'm happy with this, it feels like I'm just doing
dynamic dispatch myself. The benefit is that the BVH owns the Vec
instead of a Vec<&Instance> and you can get this type information by
matching on the enums instead of having to override various methods and
do nasty dynamic casts.