Thinking about how to separate execution environment
This will be needed so that the distributed rendering system can fit in
cleaner but I'm not happy with this current design. The way the Exec
trait is now we wouldn't be able to re-use the MultiThreaded exec since
it saves out the images when the distributed worker needs to send them
back to the master. So some more iteration on this is needed.