Write instructions asynchronously and clean up finished workers
We now also check that we've actually read the header size before
starting to read the rest of the frame data being sent.

The worker's instructions are also sent in the async event loop now and
when a worker hangs up we deregister that stream from the event loop
after shutting down the socket completely.