Convert MERL to use the arena. This required constraining 'c in bsdf
Since we want the Merl BxDF to be able to refer to the loaded file data
in the Merl Material we need to enforce that the BSDF holding the BxDF
can't outlive the material which produced it. This allows us to safely
borrow stuff from the parent material in the child BxDFs.